+++ CODEOWNERS
@@ -9,4 +9,4 @@
 # AZURE FUNCTIONS TEAM
 # For all file changes, github would automatically include the following people in the PRs.
 #
-* @vrdmr @gavin-aguiar @YunchuWang @pdthummar
+* @vrdmr @gavin-aguiar @YunchuWang @pdthummar @hallvictoria
